DataFrame.delete()
